cache friendly feature selection
This is a cache-friendlier approach. Instead of trying each feature one
at a time, we try each sample one at a time and count all the features
at once. This way, we keep hitting the same feature buffer again and
again for each sample.

In a preliminary test, this gave a ~4x speedup.